requirements.txt
Django

mongoengine

pymongo

bcrypt

django-crispy-forms